Updated Payable Protocol and updated the Savings Model to comply

+++ FinCal/Models/Protocols/Payable.swift
@@ -19,5 +19,5 @@ protocol Payable {
     var monthlyPayment: Double { get set }
 
     /// Duration of Payable in years
-    var duration: Int { get set }
+    var duration: Double { get set }
 }
